Added isLegalMove and isCapture to board class.
isLegalMove will return true if user's clicked square is a possible movement, isCapture returns true if the clicked square is a capture. Just getting basic ideas for what we will need to do come time to finalize the gui/board integration. These two do not rely on each other and by no means restrict movement to only allow for a capture if one is available, this is just for testing the gui and the movements for each piece.