Fixed some pieces
Spaced out a bit of things among the classes, just trying to make them easier to read.

Changed the pawn class so that it now checks the 0 column.

Changed the bishop and rook classes to switch the direction they check after discovering a piece in either of their paths. Doesn't break, just changes direction and continues to check.