Updated pieces and board
Made a clear method for the pieces, it is called when you click on a piece before the piece checks its movements. 

Clear deletes everything in the movement array, and is called in the controller before checking for legal moves. 

Within checking for legal moves, the movement method is called for the selected piece, so the pieces still work, but they now update their available moves with every click.