GUI is hopefully prepared for the game's undo.
We just need the Board (or a Game class) to handle undoing on behalf of
Pieces.  The Board must preserve the original Piece instances somewhere,
since the Meshes are bound to properties of those Pieces.  So when a
piece is killed, the Board must store the killed piece somewhere just in
case the user presses Undo.