Rudimentary previous state board.
Does not work with undo/redo.  Seems to work for
movements forward in time, since the "redo" function
does not rely on the pointers to the pieces in a
delta.  Currently, deltas are attached to the pieces
in the big board, so we must accomidate to make the
same deltas work on the miniboard and thereby make
undo/redo possible for the miniboard.